Refactor and decouple include directory for major refactor ahead
This commit updates the directory structure for the include
directory. This is a major step before the actual refactoring of
code that will occur next.
