+++ include/charmlite/core/core.hpp
@@ -1,7 +1,7 @@
 #ifndef __CMK_CORE_HH__
 #define __CMK_CORE_HH__
 
-#include "common.hpp"
+#include <charmlite/core/common.hpp>
 #if CHARMLITE_TOPOLOGY
 #include "TopoManager.h"
 #include "spanningTree.h"
