+++ include/charmlite/utilities/math.hpp
